# Switch FixedSizeReservoir to use math/rand/v2