Update FixedSizeReservoir to no longer allocate a RNG and instead use the auto seeded math/rand/v2 functions.

Benchmark:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/metric/exemplar
cpu: Apple M1 Max
                               │   old.txt    │               new.txt                │
                               │    sec/op    │    sec/op     vs base                │
AllocsNewFixedSizeReservoir-10   12.327µ ± 6%   1.298µ ± 14%  -89.47% (p=0.000 n=10)

                               │   old.txt    │               new.txt                │
                               │     B/op     │     B/op      vs base                │
AllocsNewFixedSizeReservoir-10   19.37Ki ± 0%   14.05Ki ± 0%  -27.43% (p=0.000 n=10)

                               │  old.txt   │              new.txt               │
                               │ allocs/op  │ allocs/op   vs base                │
AllocsNewFixedSizeReservoir-10   5.000 ± 0%   3.000 ± 0%  -40.00% (p=0.000 n=10)
```

Fixes #6260.