@pkwarren 2025-04-24
> Are there any benchmarks proving that this reduces the memory usage? Can you add a [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) result in the PR description?

Would you like for me to add a benchmark with this PR or just create one locally to include in the PR description?

https://cs.opensource.google/go/go/+/refs/tags/go1.24.2:src/math/rand/rng.go;l=180-184;drc=d8d22664c65b0054fc9ea6ba7e945481348fc8fc is the rngSource struct. The total size will be 16 bytes for ints (64-bit) and a 4856 (607*8) byte vec array, so each takes up 4872 bytes. Ref: https://go.dev/play/p/JrhIla1atjr