@pellared 2025-04-24
> Would you like for me to add a benchmark with this PR or just create one locally to include in the PR description?

First check if the existing benchmarks as they may already showcase the improvement. If there are none then please add a benchmark as part of this PR.