@pelageech 2025-04-25
Despite the zero number of allocations, global rand/v2 starts losing performance with a high level of parallelism
https://go.dev/play/p/R2SyyUIdZD_H
```
goos: darwin
goarch: arm64
pkg: randBench
cpu: Apple M3 Pro
BenchmarkRandom
BenchmarkRandom/iters=100
BenchmarkRandom/iters=100/rand=V1_Local
BenchmarkRandom/iters=100/rand=V1_Local-12         	  653746	      1918 ns/op	    5376 B/op	       1 allocs/op
BenchmarkRandom/iters=100/rand=V2_Global
BenchmarkRandom/iters=100/rand=V2_Global-12        	22497363	        53.32 ns/op	       0 B/op	       0 allocs/op
BenchmarkRandom/iters=100/rand=V2_Local
BenchmarkRandom/iters=100/rand=V2_Local-12         	34906467	        32.93 ns/op	      16 B/op	       1 allocs/op
BenchmarkRandom/iters=10000
BenchmarkRandom/iters=10000/rand=V1_Local
BenchmarkRandom/iters=10000/rand=V1_Local-12       	  280903	      4168 ns/op	    5376 B/op	       1 allocs/op
BenchmarkRandom/iters=10000/rand=V2_Global
BenchmarkRandom/iters=10000/rand=V2_Global-12      	  226537	      5454 ns/op	       0 B/op	       0 allocs/op
BenchmarkRandom/iters=10000/rand=V2_Local
BenchmarkRandom/iters=10000/rand=V2_Local-12       	  416652	      2796 ns/op	      16 B/op	       1 allocs/op
PASS
```