@pkwarren 2025-04-25
> Despite the zero number of allocations, global rand/v2 starts losing performance with a high level of parallelism

I don't show as significant of a perf difference but I'm on an older M1. Added ChaCha8 for comparison too:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/metric/exemplar
cpu: Apple M1 Max
BenchmarkRandom
BenchmarkRandom/iters=100
BenchmarkRandom/iters=100/rand=V1_Local
BenchmarkRandom/iters=100/rand=V1_Local-10         	  405732	      2512 ns/op	    5376 B/op	       1 allocs/op
BenchmarkRandom/iters=100/rand=V2_Global
BenchmarkRandom/iters=100/rand=V2_Global-10        	18214267	        62.92 ns/op	       0 B/op	       0 allocs/op
BenchmarkRandom/iters=100/rand=V2_Local_PCG
BenchmarkRandom/iters=100/rand=V2_Local_PCG-10     	21416030	        54.96 ns/op	      16 B/op	       1 allocs/op
BenchmarkRandom/iters=100/rand=V2_Local_ChaCha8
BenchmarkRandom/iters=100/rand=V2_Local_ChaCha8-10 	 4503835	       232.7 ns/op	     496 B/op	       8 allocs/op
BenchmarkRandom/iters=10000
BenchmarkRandom/iters=10000/rand=V1_Local
BenchmarkRandom/iters=10000/rand=V1_Local-10       	  217724	      5320 ns/op	    5376 B/op	       1 allocs/op
BenchmarkRandom/iters=10000/rand=V2_Global
BenchmarkRandom/iters=10000/rand=V2_Global-10      	  196062	      6031 ns/op	       0 B/op	       0 allocs/op
BenchmarkRandom/iters=10000/rand=V2_Local_PCG
BenchmarkRandom/iters=10000/rand=V2_Local_PCG-10   	  233930	      5088 ns/op	      16 B/op	       1 allocs/op
BenchmarkRandom/iters=10000/rand=V2_Local_ChaCha8
BenchmarkRandom/iters=10000/rand=V2_Local_ChaCha8-10         	  201134	      5835 ns/op	     496 B/op	       8 allocs/op
```

If you'd prefer though and PCG is sufficient for OTEL's needs, I could switch that out. 16 bytes is a lot better than 5376.