@pkwarren 2025-04-25
Ignoring allocations, here's how they stack up perf wise on my machine (with 10k iterations):
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/metric/exemplar
cpu: Apple M1 Max
BenchmarkRandom
BenchmarkRandom/rand=V1_Local
BenchmarkRandom/rand=V1_Local-10         	  354750	      3052 ns/op
BenchmarkRandom/rand=V2_Global
BenchmarkRandom/rand=V2_Global-10        	  191882	      6313 ns/op
BenchmarkRandom/rand=V2_Local_PCG
BenchmarkRandom/rand=V2_Local_PCG-10     	  216393	      5656 ns/op
BenchmarkRandom/rand=V2_Local_ChaCha8
BenchmarkRandom/rand=V2_Local_ChaCha8-10 	  172502	      6075 ns/op
```

So v1 is the fastest, but v2 global, PCG, and ChaCha8 are all fairly close.

The memory usage w/ 10k metrics however is:
* V1_Local: 46.46 MiB (4872 bytes for source)
* V2_Global: 0
* V2_Local_PCG: 0.15 MiB (16 bytes for source)
* V2_Local_ChaCha8: 3.05 MiB (320 bytes for source)

The memory calculation may be a bit off since we're not including the size of the Rand struct itself - just the source. I think either global or PCG is worth it here - I don't expect RNG time to be the limiting factor here and there will be CPU overhead for GC that we'd also need to account for with the ones that allocate.