@jhump 2025-04-25
> Despite the zero number of allocations, global rand/v2 starts losing performance with a high level of parallelism
https://go.dev/play/p/R2SyyUIdZD_H

What do you mean by parallelism? That benchmark doesn't show different levels of parallelism since the thing that varies is how many numbers to generate per trial (which has nothing to do with goroutines/parallelism/etc).

It's showing exactly what I'd expect: if you do 100x more work (100->10000 iterations), the RNGs take 100x longer. The reason the v1 doesn't take 100x longer is because there is so much overhead in its first iteration, which is getting amortized out with more and more iterations.

@pkwarren's benchmark above is more clear for understanding the perf of each RNG. Also, I think a benchmark of the _otel code_ that uses the RNG (not just the RNG itself) is more useful, to see if the couple of extra nanoseconds per random number is meaningful or not.