@dmathieu 2025-05-05
See also https://github.com/open-telemetry/opentelemetry-go/pull/6732
While that PR comes second and is therefore technically a duplicate, it handles the upgrade in the entire repo, so I'd be enclined to use it rather than this one.