@jhump 2025-05-05
> I did #6732 as a more or less mechanical upgrade.

@seankhliao, but it looks like you applied the same change as in this PR, to use the package-level functions for this reservoir instead of needing an RNG allocated per instance. (I left a comment over in that PR, too.)