@pkwarren 2025-05-05
> See also #6732 While that PR comes second and is therefore technically a duplicate, it handles the upgrade in the entire repo, so I'd be enclined to use it rather than this one.

I'd have been happy to make further changes but scoped it just to the area in question. I can close this out though if desired.